backend/requirements-airflow.txt
# Core Data
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2

# Data Validation
pandera==0.20.4

# Configuration & Reports
PyYAML==6.0.1
plotly==5.22.0
